Vision_Transformer_Salmon/Heroku Flask/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.0
Flask==2.2.2
future==0.18.2
gunicorn==20.1.0
huggingface-hub==0.10.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
tokenizers==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
